доработала слайдер
