+++ js/main.js
@@ -3,4 +3,4 @@ import './popup.js';
 import './actions-with-forms.js';
 import './form-validation.js';
 import './map-activation.js';
-import './actions-with-slider.js';
+import './slider-creation.js';
